+++ ts-declarations/xcode/index.d.ts
@@ -1,3 +1,19 @@
+interface pbxFile {
+  basename: string;
+  lastKnownFileType?: string;
+  group?: string;
+  path?: string;
+  fileEncoding?: number;
+  defaultEncoding?: number;
+  sourceTree: string;
+  includeInIndex?: number;
+  explicitFileType?: unknown;
+  settings?: object;
+  uuid?: string;
+  fileRef: string;
+  target?: string;
+}
+
 declare module 'xcode' {
   /**
    * UUID that is a key to each fragment of PBXProject.
@@ -9,37 +25,51 @@ declare module 'xcode' {
    */
   type UUIDComment = string;
 
-  interface PBXFile {
-    basename: string;
-    lastKnownFileType?: string;
-    group?: string;
-    path?: string;
-    fileEncoding?: number;
-    defaultEncoding?: number;
-    sourceTree: string;
-    includeInIndex: number;
-    explicitFileType?: unknown;
-    settings?: object;
-    uuid?: string;
-    fileRef: string;
-    target?: string;
-  }
+  type XCObjectType =
+    | 'PBXBuildFile'
+    | 'PBXFileReference'
+    | 'PBXFrameworksBuildPhase'
+    | 'PBXGroup'
+    | 'PBXNativeTarget'
+    | 'PBXProject'
+    | 'PBXResourcesBuildPhase'
+    | 'PBXShellScriptBuildPhase'
+    | 'PBXSourcesBuildPhase'
+    | 'PBXVariantGroup'
+    | 'XCBuildConfiguration'
+    | 'XCConfigurationList';
+
+  type PBXFile = pbxFile;
 
   interface PBXProject {
     isa: 'PBXProject';
-    attributes: object;
+    attributes: {
+      LastUpgradeCheck: number;
+      TargetAttributes: Record<
+        UUID,
+        {
+          CreatedOnToolsVersion?: string;
+          TestTargetID?: UUID;
+          LastSwiftMigration?: number;
+          ProvisioningStyle?: string;
+        } & Record<string, undefined | number | string>
+      >;
+    };
     buildConfigurationList: UUID;
     buildConfigurationList_comment: string;
     compatibilityVersion: string;
     developmentRegion: string;
     hasScannedForEncodings: number;
-    knownRegions: unknown[];
+    knownRegions: string[];
     mainGroup: UUID;
     productRefGroup: UUID;
     productRefGroup_comment: string;
     projectDirPath: string;
     projectRoot: string;
-    targets: unknown[];
+    targets: {
+      value: UUID;
+      comment: string;
+    }[];
   }
 
   interface PBXNativeTarget {
@@ -57,6 +87,63 @@ declare module 'xcode' {
     productReference: UUID;
     productReference_comment: string;
     productType: string;
+  }
+
+  interface PBXBuildFile {
+    isa: 'PBXBuildFile';
+    fileRef: UUID;
+    // "AppDelegate.m"
+    fileRef_comment: string;
+  }
+
+  interface XCConfigurationList {
+    isa: 'XCConfigurationList';
+    buildConfigurations: {
+      value: UUID;
+      comment: string | 'Release' | 'Debug';
+    }[];
+    defaultConfigurationIsVisible: number;
+    defaultConfigurationName: string;
+  }
+
+  interface XCBuildConfiguration {
+    isa: 'XCBuildConfiguration';
+    baseConfigurationReference: UUID;
+    baseConfigurationReference_comment: string;
+    buildSettings: Record<string, string | undefined | number | unknown[]> & {
+      // '"$(TARGET_NAME)"',
+      PRODUCT_NAME?: string;
+      // '"io.expo.demo.$(PRODUCT_NAME:rfc1034identifier)"',
+      PRODUCT_BUNDLE_IDENTIFIER?: string;
+      PROVISIONING_PROFILE_SPECIFIER?: string;
+      // '"$(BUILT_PRODUCTS_DIR)/rni.app/rni"'
+      TEST_HOST?: any;
+      DEVELOPMENT_TEAM?: string;
+      CODE_SIGN_IDENTITY?: string;
+      CODE_SIGN_STYLE?: string;
+      // '"$(TEST_HOST)"'
+      BUNDLE_LOADER?: string;
+      GCC_PREPROCESSOR_DEFINITIONS?: unknown[];
+      INFOPLIST_FILE?: string;
+      IPHONEOS_DEPLOYMENT_TARGET?: string;
+      LD_RUNPATH_SEARCH_PATHS?: string;
+      OTHER_LDFLAGS?: unknown[];
+      ASSETCATALOG_COMPILER_APPICON_NAME?: string;
+      ASSETCATALOG_COMPILER_LAUNCHIMAGE_NAME?: string;
+      CLANG_ANALYZER_NONNULL?: string;
+      CLANG_WARN_DOCUMENTATION_COMMENTS?: string;
+      CLANG_WARN_INFINITE_RECURSION?: string;
+      CLANG_WARN_SUSPICIOUS_MOVE?: string;
+      DEBUG_INFORMATION_FORMAT?: string;
+      ENABLE_TESTABILITY?: string;
+      GCC_NO_COMMON_BLOCKS?: string;
+      // 'appletvos'
+      SDKROOT?: string;
+      TARGETED_DEVICE_FAMILY?: number | string;
+      // '10.0'
+      TVOS_DEPLOYMENT_TARGET?: string;
+    };
+    name: string;
   }
 
   type ProductType =
@@ -84,7 +171,7 @@ declare module 'xcode' {
     sourceTree: '"<group>"' | unknown;
   }
 
-  export class project {
+  export class XcodeProject {
     constructor(pbxprojPath: string);
 
     /**
@@ -92,9 +179,27 @@ declare module 'xcode' {
      */
     filepath: string;
 
+    // Ex: '$(TARGET_NAME)'
+    productName: string;
+
     hash: {
       headComment: string;
-      project: unknown;
+      project: {
+        archiveVersion: number;
+        objectVersion: number;
+        objects: {
+          [T in XCObjectType]: Record<
+            string,
+            {
+              isa: T;
+              name: string;
+              [key: string]: any;
+            }
+          >;
+        };
+        rootObject: string;
+        rootObject_comment: string;
+      };
     };
 
     // ------------------------------------------------------------------------
@@ -107,6 +212,8 @@ declare module 'xcode' {
      * First step to be executed while working with `.pbxproj` file.
      */
     parse(callback?: (err: Error | null, results?: string) => void): this;
+
+    parseSync(): void;
 
     /**
      * @returns Content of .pbxproj file.
@@ -148,7 +255,7 @@ declare module 'xcode' {
     removePbxGroup(groupName: unknown): void;
     addToPbxProjectSection(target: unknown): void;
     addToPbxNativeTargetSection(target: unknown): void;
-    addToPbxFileReferenceSection(file: unknown): void;
+    addToPbxFileReferenceSection(file: any): void;
     removeFromPbxFileReferenceSection(file: unknown): unknown;
     addToXcVersionGroupSection(file: unknown): void;
     addToPluginsPbxGroup(file: unknown): void;
@@ -210,30 +317,31 @@ declare module 'xcode' {
     /**
      * Retrieves main part describing PBXProjects that are available in `.pbxproj` file.
      */
-    pbxProjectSection(): { [key: UUID]: PBXProject };
-    pbxBuildFileSection(): unknown;
-    pbxXCBuildConfigurationSection(): unknown;
+    pbxProjectSection(): Record<UUID, PBXProject> & Record<UUIDComment, string>;
+    pbxBuildFileSection(): Record<UUID, PBXBuildFile> & Record<UUIDComment, string>;
+    pbxXCBuildConfigurationSection(): Record<UUID, XCBuildConfiguration> &
+      Record<UUIDComment, string>;
     pbxFileReferenceSection(): Record<UUID, PBXFile> & Record<UUIDComment, string>;
-    pbxNativeTargetSection(): unknown;
+    pbxNativeTargetSection(): Record<UUID, PBXNativeTarget> & Record<UUIDComment, string>;
     xcVersionGroupSection(): unknown;
-    pbxXCConfigurationList(): unknown;
-    pbxGroupByName(name: unknown): unknown;
+    pbxXCConfigurationList(): Record<UUID, XCConfigurationList> & Record<UUIDComment, string>;
+    pbxGroupByName(name: string): PBXGroup | undefined;
     /**
      * @param targetName in most cases it's the name of the application
      */
     pbxTargetByName(targetName: string): PBXNativeTarget | undefined;
-    findTargetKey(name: unknown): string;
-    pbxItemByComment(name: unknown, pbxSectionName: unknown): unknown;
+    findTargetKey(name: string): string;
+    pbxItemByComment(name: string, pbxSectionName: unknown): unknown;
     pbxSourcesBuildPhaseObj(target: unknown): unknown;
     pbxResourcesBuildPhaseObj(target: unknown): unknown;
     pbxFrameworksBuildPhaseObj(target: unknown): unknown;
     pbxEmbedFrameworksBuildPhaseObj(target: unknown): unknown;
     buildPhase(group: unknown, target: unknown): string;
-    buildPhaseObject(name: unknown, group: unknown, target: unknown): unknown;
+    buildPhaseObject(name: string, group: unknown, target: unknown): unknown;
     addBuildProperty(prop: unknown, value: unknown, build_name: unknown): void;
     removeBuildProperty(prop: unknown, build_name: unknown): void;
     updateBuildProperty(prop: string, value: unknown, build: string): void;
-    updateProductName(name: unknown): void;
+    updateProductName(name: string): void;
     removeFromFrameworkSearchPaths(file: unknown): void;
     addToFrameworkSearchPaths(file: unknown): void;
     removeFromLibrarySearchPaths(file: unknown): void;
@@ -271,7 +379,7 @@ declare module 'xcode' {
      */
     getFirstProject(): { uuid: UUID; firstProject: PBXProject };
     getFirstTarget(): {
-      uuid: unknown;
+      uuid: UUID;
       firstTarget: unknown;
     };
     /**
@@ -283,7 +391,7 @@ declare module 'xcode' {
     addToPbxGroup(file: PBXFile, groupKey: UUID): void;
     pbxCreateGroupWithType(name: unknown, pathName: unknown, groupType: unknown): unknown;
     pbxCreateVariantGroup(name: unknown): unknown;
-    pbxCreateGroup(name: unknown, pathName: unknown): unknown;
+    pbxCreateGroup(name: string, pathName: string): UUID;
     removeFromPbxGroupAndType(file: unknown, groupKey: unknown, groupType: unknown): void;
     removeFromPbxGroup(file: unknown, groupKey: unknown): void;
     removeFromPbxVariantGroup(file: unknown, groupKey: unknown): void;
@@ -306,10 +414,10 @@ declare module 'xcode' {
       fileRef: unknown;
       basename: unknown;
     };
-    addKnownRegion(name: unknown): void;
-    removeKnownRegion(name: unknown): void;
-    hasKnownRegion(name: unknown): boolean;
-    getPBXObject(name: unknown): unknown;
+    addKnownRegion(name: string): void;
+    removeKnownRegion(name: string): void;
+    hasKnownRegion(name: string): boolean;
+    getPBXObject(name: string): unknown;
     /**
      * - creates `PBXFile`
      * - adds to `PBXFileReference` section
@@ -339,5 +447,26 @@ declare module 'xcode' {
     addDataModelDocument(filePath: unknown, group: unknown, opt: unknown): unknown;
     addTargetAttribute(prop: unknown, value: unknown, target: unknown): void;
     removeTargetAttribute(prop: unknown, target: unknown): void;
+  }
+
+  export function project(projectPath: string): XcodeProject;
+}
+
+declare module 'xcode/lib/pbxFile' {
+  export default class PBXFile implements pbxFile {
+    constructor(file: string);
+    basename: string;
+    lastKnownFileType?: string;
+    group?: string;
+    path?: string;
+    fileEncoding?: number;
+    defaultEncoding?: number;
+    sourceTree: string;
+    includeInIndex?: number;
+    explicitFileType?: unknown;
+    settings?: object;
+    uuid?: string;
+    fileRef: string;
+    target?: string;
   }
 }